+++ newton/_src/solvers/xpbd/solver_xpbd.py
@@ -836,8 +836,6 @@ class SolverXPBD(SolverBase, CouplingInterface):
                             contacts.rigid_contact_point1,
                             contacts.rigid_contact_offset0,
                             contacts.rigid_contact_offset1,
-                            contacts.rigid_contact_margin0,
-                            contacts.rigid_contact_margin1,
                             rigid_contact_inv_weight_init,
                             model.gravity,
                             dt,